Agentic-AI/requirements.txt
langchain
python-dotenv
langchain_google_genai
google-generativeai
ipykernel
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
chromadb
langchain-community
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit